Refactor opcode execution
This commit moves all of the opcode methods in the Cpu struct out of the
Cpu struct and into their own file. They instead borrow a Cpu struct and
act on it. There is now an INSTRUCTIONS array of function pointers of
size 256. The execute method uses the opcode to index the array and get
the function. This removes the need for the massive match statement in
the execute method.